backend/requirements.txt
Django==5.0.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
django-filter==23.5
Pillow==10.2.0
python-dotenv==1.0.1
psycopg2-binary==2.9.9  # For PostgreSQL support
gunicorn==21.2.0  # For production deployment 